delete unused files and moves around some constants
